Document icecream contract (and remove v0 version)
