+++ examples/icecream/Cargo.toml
@@ -8,7 +8,7 @@ license = "MPL-2.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-concordium-std = "1"
+concordium-std = {path = "../../concordium-std"}
 
 [lib]
 crate-type=["cdylib", "rlib"]
