--- examples/icecream1/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "icecream1"
-version = "0.1.0"
-authors = ["Concordium <developers@concordium.com>"]
-edition = "2018"
-license = "MPL-2.0"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-concordium-std = {path = "../../concordium-std"}
-
-[lib]
-crate-type=["cdylib", "rlib"]
